build: add new Misc Tests CI target
We have `tree.TestPrettyData` and
`metamorphic.TestPebbleEquivalenceNightly` tests that have `nightly` Go
build tag set, but we don't have any CI targets that would run this.
(The closest one is Lint Nightly but it's specific to TestNightlyLint.)

This commit adds a couple of scripts for the new CI target aimed to run
miscellaneous tests on a nightly basis. These scripts were inspired by
the `optimizer_tests*` ones. Two tests mentioned are refactored to no
longer use `go:build nightly` tag in favor of COCKROACH_MISC_NIGHTLY env
var, which if not set or is `false`, skips the test.

Additionally, we reduce the op count for the pebble test from 1M to 500k
to speed it up a bit (with the reduced count it takes a bit over an
hour, so we give it a conservative 3h timeout).

Release note: None
