+++ pkg/sql/sem/tree/pretty_test.go
@@ -22,6 +22,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/types"
 	"github.com/cockroachdb/cockroach/pkg/testutils/datapathutils"
+	"github.com/cockroachdb/cockroach/pkg/testutils/skip"
 	"github.com/cockroachdb/cockroach/pkg/testutils/sqlutils"
 	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
@@ -39,11 +40,13 @@ var (
 	}()
 )
 
-// TestPrettyDataShort reads in a single SQL statement from a file, formats
-// it at 40 characters width, and compares that output to a known-good
-// output file. It is most useful when changing or implementing the
-// doc interface for a node, and should be used to compare and verify
-// the changed output.
+// TestPrettyDataShort reads in a single SQL statement from a file, formats it
+// at 40 characters width, and compares that output to a known-good output file.
+// It is most useful when changing or implementing the doc interface for a node,
+// and should be used to compare and verify the changed output.
+//
+// Unlike TestPrettyData, this test only formats the statements at a single
+// width.
 func TestPrettyDataShort(t *testing.T) {
 	defer leaktest.AfterTest(t)()
 	defer log.Scope(t).Close(t)
@@ -57,15 +60,47 @@ func TestPrettyDataShort(t *testing.T) {
 	cfg := testPrettyCfg
 	cfg.Align = tree.PrettyNoAlign
 	t.Run("ref", func(t *testing.T) {
-		runTestPrettyData(t, "ref", cfg, matches, true /*short*/)
+		runTestPrettyData(t, "ref", cfg, matches, true /* short */)
 	})
 	cfg.Align = tree.PrettyAlignAndDeindent
 	t.Run("align-deindent", func(t *testing.T) {
-		runTestPrettyData(t, "align-deindent", cfg, matches, true /*short*/)
+		runTestPrettyData(t, "align-deindent", cfg, matches, true /* short */)
 	})
 	cfg.Align = tree.PrettyAlignOnly
 	t.Run("align-only", func(t *testing.T) {
-		runTestPrettyData(t, "align-only", cfg, matches, true /*short*/)
+		runTestPrettyData(t, "align-only", cfg, matches, true /* short */)
+	})
+}
+
+// TestPrettyData reads in a single SQL statement from a file, formats it at all
+// line lengths, and compares that output to a known-good output file. It is
+// most useful when changing or implementing the doc interface for a node, and
+// should be used to compare and verify the changed output.
+//
+// Unlike TestPrettyDataShort, this test formats the statement at all possible
+// widths (based on the length of the statement itself).
+func TestPrettyData(t *testing.T) {
+	defer leaktest.AfterTest(t)()
+	defer log.Scope(t).Close(t)
+
+	skip.IfNotMiscNightly(t)
+
+	matches, err := filepath.Glob(datapathutils.TestDataPath(t, "pretty", "*.sql"))
+	if err != nil {
+		t.Fatal(err)
+	}
+	cfg := testPrettyCfg
+	cfg.Align = tree.PrettyNoAlign
+	t.Run("ref", func(t *testing.T) {
+		runTestPrettyData(t, "ref", cfg, matches, false /* short */)
+	})
+	cfg.Align = tree.PrettyAlignAndDeindent
+	t.Run("align-deindent", func(t *testing.T) {
+		runTestPrettyData(t, "align-deindent", cfg, matches, false /* short */)
+	})
+	cfg.Align = tree.PrettyAlignOnly
+	t.Run("align-only", func(t *testing.T) {
+		runTestPrettyData(t, "align-only", cfg, matches, false /* short */)
 	})
 }
