+++ pkg/storage/metamorphic/meta_test.go
@@ -167,13 +167,34 @@ func TestPebbleEquivalence(t *testing.T) {
 	defer leaktest.AfterTest(t)()
 	defer log.Scope(t).Close(t)
 
+	// This test times out with the race detector enabled.
 	skip.UnderRace(t)
+	runPebbleEquivalenceTest(t)
+}
+
+func TestPebbleEquivalenceNightly(t *testing.T) {
+	defer leaktest.AfterTest(t)()
+	defer log.Scope(t).Close(t)
+
+	skip.IfNotMiscNightly(t)
+
+	// This test times out with the race detector enabled.
+	skip.UnderRace(t)
+	if *opCount < 500000 {
+		oldOpCount := *opCount
+		// Override number of operations to at least half a million.
+		*opCount = 500000
+
+		defer func() {
+			*opCount = oldOpCount
+		}()
+	}
+
 	runPebbleEquivalenceTest(t)
 }
 
 func runPebbleEquivalenceTest(t *testing.T) {
 	ctx := context.Background()
-	// This test times out with the race detector enabled.
 	_, seed := randutil.NewTestRand()
 
 	engineSeqs := make([]engineSequence, 0, numStandardOptions+numRandomOptions)
