+++ pkg/testutils/skip/skip.go
@@ -245,3 +245,15 @@ func maybeSkip(t SkippableTest, reason string, args ...interface{}) {
 
 	t.Skip(append([]interface{}{reason}, args...)...)
 }
+
+var miscNightly = envutil.EnvOrDefaultBool("COCKROACH_MISC_NIGHTLY", false)
+
+// IfNotMiscNightly skips this test unless the COCKROACH_MISC_NIGHTLY env var is
+// set to 'true'.
+//
+// Does not respect COCKROACH_FORCE_RUN_SKIPPED_TESTS.
+func IfNotMiscNightly(t SkippableTest) {
+	if !miscNightly {
+		t.Skip("only runs in Misc Nightly CI")
+	}
+}
